Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
mysql-connector-python
alembic  # optional
